Move insights into modules folder
