+++ packages/cli/src/databases/entities/index.ts
@@ -33,9 +33,9 @@ import { WorkflowEntity } from './workflow-entity';
 import { WorkflowHistory } from './workflow-history';
 import { WorkflowStatistics } from './workflow-statistics';
 import { WorkflowTagMapping } from './workflow-tag-mapping';
-import { InsightsByPeriod } from '../../insights/entities/insights-by-period';
-import { InsightsMetadata } from '../../insights/entities/insights-metadata';
-import { InsightsRaw } from '../../insights/entities/insights-raw';
+import { InsightsByPeriod } from '../../modules/insights/entities/insights-by-period';
+import { InsightsMetadata } from '../../modules/insights/entities/insights-metadata';
+import { InsightsRaw } from '../../modules/insights/entities/insights-raw';
 
 export const entities = {
 	AnnotationTagEntity,
