+++ packages/cli/src/databases/repositories/insights-by-period.repository.ts
@@ -1,7 +1,7 @@
 import { Service } from '@n8n/di';
 import { DataSource, Repository } from '@n8n/typeorm';
 
-import { InsightsByPeriod } from '../../insights/entities/insights-by-period';
+import { InsightsByPeriod } from '../../modules/insights/entities/insights-by-period';
 
 @Service()
 export class InsightsByPeriodRepository extends Repository<InsightsByPeriod> {
