+++ packages/cli/src/modules/insights/entities/__tests__/insights-by-period.test.ts
@@ -8,7 +8,7 @@ import { createMetadata, createRawInsightsEvent } from '@test-integration/db/ins
 import { createTeamProject } from '@test-integration/db/projects';
 import { createWorkflow } from '@test-integration/db/workflows';
 
-import * as testDb from '../../../../test/integration/shared/test-db';
+import * as testDb from '../../../../../test/integration/shared/test-db';
 import { InsightsByPeriod } from '../insights-by-period';
 import type { PeriodUnits, TypeUnits } from '../insights-shared';
