+++ packages/cli/src/modules/insights/entities/insights-by-period.ts
@@ -10,7 +10,7 @@ import {
 	PeriodUnitToNumber,
 	TypeToNumber,
 } from './insights-shared';
-import { datetimeColumnType } from '../../databases/entities/abstract-entity';
+import { datetimeColumnType } from '../../../databases/entities/abstract-entity';
 
 @Entity()
 export class InsightsByPeriod extends BaseEntity {
