+++ packages/cli/src/modules/insights/entities/insights-metadata.ts
