+++ packages/cli/src/modules/insights/entities/insights-shared.ts
