+++ packages/cli/src/modules/insights/insights.config.ts
