+++ packages/cli/src/modules/insights/insights.module.ts
