+++ packages/cli/test/integration/shared/db/insights.ts
@@ -5,9 +5,9 @@ import type { WorkflowEntity } from '@/databases/entities/workflow-entity';
 import { InsightsByPeriodRepository } from '@/databases/repositories/insights-by-period.repository';
 import { InsightsMetadataRepository } from '@/databases/repositories/insights-metadata.repository';
 import { InsightsRawRepository } from '@/databases/repositories/insights-raw.repository';
-import { InsightsByPeriod } from '@/insights/entities/insights-by-period';
-import { InsightsMetadata } from '@/insights/entities/insights-metadata';
-import { InsightsRaw } from '@/insights/entities/insights-raw';
+import { InsightsByPeriod } from '@/modules/insights/entities/insights-by-period';
+import { InsightsMetadata } from '@/modules/insights/entities/insights-metadata';
+import { InsightsRaw } from '@/modules/insights/entities/insights-raw';
 
 import { getWorkflowSharing } from './workflows';
